Summary page styling, timeout popup now configurable, moved signin to after landingpage
