+++ src/components/AppComponents/SummaryPage/SummaryPage.tsx
@@ -9,7 +9,7 @@ export default function SummaryPage(props){
         return (<>
                 <MainWrapper>
                     <div className='govuk-panel govuk-panel--confirmation govuk-!-margin-bottom-7'>
-                        {summaryBanner}
+                        <h1 className='govuk-panel__title'> {summaryBanner} </h1>
                     </div>
                         <ParsedHTML htmlString={summaryContent} />
                 </MainWrapper>
@@ -19,7 +19,7 @@ export default function SummaryPage(props){
     
     return  <>
         <MainWrapper>
-            <h1>
+            <h1 className='govuk-heading-l'>
                 {summaryTitle}
             </h1>
             <ParsedHTML htmlString={summaryContent} />
