+++ src/components/BaseComponents/MainWrapper/index.tsx
@@ -18,7 +18,7 @@ export default function MainWrapper ({children}) {
                   className='govuk-link hmrc-report-technical-issue '
                   rel='noreferrer noopener'
                   target='_blank'
-                  href={`${hmrcURL}contact/report-technical-problem?newTab=true&service=463&referrerUrl=${referrerURL}`}
+                  href={`${hmrcURL}contact/report-technical-problem?newTab=true&service=463&referrerUrl=${window.location}`}
                 >
                   {t('PAGE_NOT_WORKING_PROPERLY')} {t("OPENS_IN_NEW_TAB")}
                 </a>
